+++ src/include/LSP/Workspace.hpp
@@ -6,6 +6,7 @@
 #include "Protocol/LanguageFeatures.hpp"
 #include "Protocol/SignatureHelp.hpp"
 #include "Protocol/SemanticTokens.hpp"
+#include "Protocol/Extensions.hpp"
 #include "LSP/Client.hpp"
 #include "LSP/WorkspaceFileResolver.hpp"
 #include "LSP/LuauExt.hpp"
@@ -118,6 +119,8 @@ public:
     std::optional<std::vector<lsp::DocumentSymbol>> documentSymbol(const lsp::DocumentSymbolParams& params);
     std::optional<std::vector<lsp::WorkspaceSymbol>> workspaceSymbol(const lsp::WorkspaceSymbolParams& params);
     std::optional<lsp::SemanticTokens> semanticTokens(const lsp::SemanticTokensParams& params);
+
+    lsp::CompilerRemarksResult compilerRemarks(const lsp::CompilerRemarksParams& params);
 
     bool updateSourceMap();